Improves daily bar retrieval performance
Refactors the `GetDailyBarsAsync` method in the `QuoteRepository`
to use `ExecuteReaderAsync` for improved performance. This change
streamlines the data retrieval process and reduces overhead.

Also updates the `DailyBar` model to use `float` and `long` types
for better consistency and performance.

+++ Core/ML/TimeSeriesUtils.cs
@@ -26,11 +26,11 @@ namespace Core.ML
     public class DailyBar
     {
         public DateTime Date { get; set; }
-        public decimal Open { get; set; }
-        public decimal High { get; set; }
-        public decimal Low { get; set; }
-        public decimal Close { get; set; }
-        public decimal Volume { get; set; }
+        public float Open { get; set; }
+        public float High { get; set; }
+        public float Low { get; set; }
+        public float Close { get; set; }
+        public long Volume { get; set; }
     }
 
     public class FeatureRow
